Ensures factor bundle correctly matches entries to output streams
The documentation for [`_.values(object)`](https://lodash.com/docs#values) gives an example:

>     _.values(new Foo);
>     // → [1, 2] (iteration order is not guaranteed)

Since "iteration order is not guaranteed", the array of streams passed to factor-bundle may not be in the same order as the corresponding files in `entries`.  Factor bundle [depends on this correlation](https://github.com/substack/factor-bundle/blob/master/index.js#L73) when building its pipeline.

This commit ensures that there is never a mismatch.